Converted all the promise functions in levelSandbox.js to one object.   This way we can export just 1 object instead of individual functions.
 Also made alot of changes to simpleChain.js.  Implemented getBlock(), getBlockHeight(), addBlock().

 getBlock() returns a promise which will then be used to get a block at a specific height.
getBlockHeight() gets the latest block height.  This is important since we won't be able to get the actual block without knowing the blockheight.  The addBlock() method will be used for adding the a block to the blockchain.  The issue is this method is a bit large

Also added a method called addSingleBlock which is used to decrease the size of addBlock()