Did alot of work in regards to the MemPoolController And everyhting should be working as expected.
The only issue that came up was the verifyAddressRequest() method.   In a nutshell this method has access to the validMemPoolEntries map.  But Adding a block belonged to the BlockController.    Since I couldn't find an easy way to make a call between controllers, I decided to unify both controllers under the umbrella of the StarRegistryController. The goal of doing this was to keep things simple.   If this project continues then common functions/objects would have to be placed in a ulity section of some kind.  Anyway, the StarRegistryController will be added as well and the MemPool/Blockhain controllers will be removed