Added the getBlockByWalletAddress() for the API/simpleChain/levelsandbox.  This is essentially a way to search the blockchain by wallet address.  Also added some error handling for both getBlockByWalletAddress() and getBlockByHash()
When debugging the leveldb part of the code, I realized that I had old blocks that didn't contain star data and the search was failing since it couldn't find the btc address within the newly formed body{}.   So the blockchain had to be recreated and with relevant data.   I decided to use the sun as the Genesis block and found RA/DEC data from my phone.  Also i need to add test data so I can easily add stars to this blockchain.   Also unit testing would help